chore: priming-roll test extraction, dead-branch removal, citation fixes
- Extract the priming-roll unit tests (both describe blocks) out of the
  208KB wrap-pipeline.test.js into a dedicated test/wrap-step-priming-roll.test.js,
  matching every sibling wrap-step's convention (advances TST-4X8N). The
  pipeline-level ordering/commit/drawer tests that reference the step stay.
- Remove the now-unreachable "no ### Chunk N:" fallback branch of
  _renderPointerBody + its direct test: since #515 made run() skip a
  chunk-less plan before rendering, _selectPointer can only yield an
  allDone or current pointer, never an all-null one.
- Repoint two provenance citations (lib/continuity.js, FEATURES.md) to
  the archived continuity-contract.md path.
- Correct a stale premise comment on the commit/drawer defensive-render
  test (post-#515 that staged shape is unreachable); file WRP-6C4M to
  reconcile the parallel path.

No behavior change.
