+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to TangleClaw are documented in this file.
 
 - **The `priming-roll` wrap step (`next-session-prime`) now skips — instead of blocking the wrap — when the resolved plan has no `### Chunk N:` headings (#515).** `lib/wrap-steps/priming-roll.js` treated a chunk-less plan asymmetrically: the single-plan resolve path (`.claude/plans/` holds exactly one `.md`) returned it and let `run()` hard-**block** on zero chunks, while the multi-plan path already **skipped** such a plan via `_isPlanInProgress`. So the same spec/design doc (or a not-yet-chunked plan) failed the wrap when it stood alone but was silently dropped when it had company — the `feedback_symmetric_capability_gates` class of gate asymmetry. Observed live 2026-07-09: TC's own wrap went BLOCKED on `.claude/plans/continuity-contract.md` (a Phase-0 spec doc, zero chunks) once the real build plan was archived. The `run()` zero-chunk gate now returns `status:'skipped'` with a reason (`… no ### Chunk N: headings — nothing to roll (add chunk headings if this is meant to be an active build plan)`) and stages no priming write — honest and visible in the drawer, never halting. Applies to explicitly-configured `step.planPath`s too (a pointer at a chunk-less file is a visible skip, not a wrap failure). Re-specified the prior block-asserting test to the new contract and added regression tests pinning single-plan/multi-plan symmetry and the explicit-planPath case (`test/wrap-pipeline.test.js`).
 
+### Internal
+
+- **priming-roll hygiene: own test file, dead-branch removal, citation fixes (TST-4X8N + #515 follow-up).** Extracted the `priming-roll` unit tests (~690 lines, both `describe` blocks) out of the 208 KB `test/wrap-pipeline.test.js` into a dedicated `test/wrap-step-priming-roll.test.js`, matching every sibling wrap-step's convention (advances backlog TST-4X8N for the priming-roll slice; the pipeline-level ordering/commit/drawer tests that reference the step stay in `wrap-pipeline.test.js`). Removed the now-unreachable "no `### Chunk N:` headings" fallback branch of `_renderPointerBody` in `lib/wrap-steps/priming-roll.js` (+ its direct test): since #515 made `run()` skip a chunk-less plan before rendering, `_selectPointer` can only ever hand the renderer an `allDone` or `current` pointer, never an all-`null` one — the branch was dead code testing dead code. Repointed two provenance citations (`lib/continuity.js`, `FEATURES.md`) from `.claude/plans/continuity-contract.md` to `.claude/plans/archive/continuity-contract.md` after the spec doc was archived. No behavior change. (Left in place, noted for a future pass: the commit/drawer side carries a parallel "no parseable chunks" defensive render for the same staged shape — a separate consumer, out of scope here.)
+
 ## [4.8.0] - 2026-07-09
 
 ### Added
