+++ FEATURES.md
@@ -36,7 +36,7 @@ entries that span multiple co-equal locations.
 - **Methodologies registry** — `GET /api/methodologies` `server.js:1119`, `GET /api/methodologies/:id` `:1125`. Backend: `lib/methodologies.js:235` (`initialize`), `:614` (`listTemplates`).
 - **Sessions family** — `POST /api/sessions/:project` (launch) `:1162`, `DELETE` (kill) `:1232`, `GET /status` `:1265`, `POST /command` `:1274`, `POST /wrap` `:1304`, `POST /wrap/complete` `:1338`, `GET /peek` `:1354`, `GET /history` `:1374`. Core: `lib/sessions.js:51` (`launchSession`), `:335` (`generatePrimePrompt` docstring).
 - **Wrap pipeline runner** — `lib/wrap-pipeline.js:135` (`runWrapPipeline`).
-- **Continuity store** (Continuity Contract, `.claude/plans/continuity-contract.md`) — per-project gitignored store at `<project>/.tangleclaw/continuity/`. **Hot tier (CC-1):** curated `index.md` rewritten each wrap, read back by `generatePrimePrompt` for the "we left off at X — continue?" resume. **Warm tier (CC-2):** append-only `changelog.md` (one entry per session) + per-session `wraps/<sid>.md` 8-section summary + `search(projectPath, query, {section})` grep-over-markdown retrieval carrying the `session:<sid>` pointer. **Map (CC-3):** the index's `## Map` feature/component section, self-maintained at wrap (`updateMap` stubs touched files, prunes deleted; preserved verbatim across the rewrite) — internal hot-tier sibling of this `FEATURES.md`. **Cold tier (CC-4 / CC-4b):** uploads relocate to `sessions/<sid>/uploads/` (CC-4) and the raw session transcript snapshots to `sessions/<sid>/transcript.jsonl` + `transcript.meta.json` at wrap (CC-4b, `lib/transcript.js`). The transcript is resolved at wrap with **no hook** — a harness-adapter registry keyed by engine (`claude` implemented via a `~/.claude/projects` cwd-content match; `gemini`/`codex`/`aider`/`openclaw` are forward-compat stubs) — and scanned by `lib/secret-scan.js` (flag-only, types not values). **Operator search (CC-5, #344):** the **Session History & Search** drawer (&#128269; button on each project card → `public/history-drawer.js` + `#historyModal`). Warm global search (`searchSessions()` — filters: date range / `tags` / `type` / file-touched / `refs`; recency+match-count ranking; browse mode on empty query) → cold drill-down (`searchTranscript()` over `transcript.jsonl`, surfacing the `secretsFlagged` warning). A **Summaries / Full-transcripts** toggle switches the global box to `searchProjectTranscripts()` (`?scope=transcripts`) — greps every captured transcript in the project at once (direct transcript search, same filters, results highlighted via `<mark>`). `listSessions()` merges wrap + changelog + transcript meta per session. New schema fields feed the filters: `type:`/`[type]` from the branch prefix, `files:` from the wrap step's touched list (both forward-only; `meta.unindexed` labels the gap). Routes: `GET /api/continuity/:project/{search,sessions,sessions/:sid,sessions/:sid/transcript/search}`. All render/parse pure in `lib/continuity.js`; written by the `continuity-write` wrap step (`lib/wrap-steps/continuity-write.js`, runs after `commit`, `blocker:false`; the transcript snapshot is isolated so a copy failure never affects the warm tier). **Wrap-section depth (CC-6 → CC-8, #386):** the rendered section set resolves per-project `project.wrapSections` (CC-6 operator override) → methodology `template.json` `wrap_contract.sections` (CC-8 default, validated in `lib/methodologies.js`) → all 8 (deep fallback); `Next action` is always forced in by `effectiveWrapSections`. Prawduct omits `wrap_contract` ⇒ stays deep.
+- **Continuity store** (Continuity Contract, `.claude/plans/archive/continuity-contract.md`) — per-project gitignored store at `<project>/.tangleclaw/continuity/`. **Hot tier (CC-1):** curated `index.md` rewritten each wrap, read back by `generatePrimePrompt` for the "we left off at X — continue?" resume. **Warm tier (CC-2):** append-only `changelog.md` (one entry per session) + per-session `wraps/<sid>.md` 8-section summary + `search(projectPath, query, {section})` grep-over-markdown retrieval carrying the `session:<sid>` pointer. **Map (CC-3):** the index's `## Map` feature/component section, self-maintained at wrap (`updateMap` stubs touched files, prunes deleted; preserved verbatim across the rewrite) — internal hot-tier sibling of this `FEATURES.md`. **Cold tier (CC-4 / CC-4b):** uploads relocate to `sessions/<sid>/uploads/` (CC-4) and the raw session transcript snapshots to `sessions/<sid>/transcript.jsonl` + `transcript.meta.json` at wrap (CC-4b, `lib/transcript.js`). The transcript is resolved at wrap with **no hook** — a harness-adapter registry keyed by engine (`claude` implemented via a `~/.claude/projects` cwd-content match; `gemini`/`codex`/`aider`/`openclaw` are forward-compat stubs) — and scanned by `lib/secret-scan.js` (flag-only, types not values). **Operator search (CC-5, #344):** the **Session History & Search** drawer (&#128269; button on each project card → `public/history-drawer.js` + `#historyModal`). Warm global search (`searchSessions()` — filters: date range / `tags` / `type` / file-touched / `refs`; recency+match-count ranking; browse mode on empty query) → cold drill-down (`searchTranscript()` over `transcript.jsonl`, surfacing the `secretsFlagged` warning). A **Summaries / Full-transcripts** toggle switches the global box to `searchProjectTranscripts()` (`?scope=transcripts`) — greps every captured transcript in the project at once (direct transcript search, same filters, results highlighted via `<mark>`). `listSessions()` merges wrap + changelog + transcript meta per session. New schema fields feed the filters: `type:`/`[type]` from the branch prefix, `files:` from the wrap step's touched list (both forward-only; `meta.unindexed` labels the gap). Routes: `GET /api/continuity/:project/{search,sessions,sessions/:sid,sessions/:sid/transcript/search}`. All render/parse pure in `lib/continuity.js`; written by the `continuity-write` wrap step (`lib/wrap-steps/continuity-write.js`, runs after `commit`, `blocker:false`; the transcript snapshot is isolated so a copy failure never affects the warm tier). **Wrap-section depth (CC-6 → CC-8, #386):** the rendered section set resolves per-project `project.wrapSections` (CC-6 operator override) → methodology `template.json` `wrap_contract.sections` (CC-8 default, validated in `lib/methodologies.js`) → all 8 (deep fallback); `Next action` is always forced in by `effectiveWrapSections`. Prawduct omits `wrap_contract` ⇒ stays deep.
 - **Server-info endpoint** (#199, #235) — `GET /api/server-info` `server.js:324` (now includes `restartMechanism`). Backend: `lib/server-info.js`.
 - **Server-restart endpoint** (#235) — `POST /api/server/restart` (next to `/api/server-info`). 202 Accepted before exec; 501 when no mechanism. Backend: `lib/server-info.js#detectRestartMechanism`, `#buildRestartCommand`.
 - **PortHub leasing** — `POST /api/ports/lease` `:805`, `POST /api/ports/release` `:834`, `POST /api/ports/heartbeat` `:843`, `GET /api/ports` `:781`, `POST /api/ports/sync` `:828`. Backend: `lib/porthub.js`.
