+++ lib/continuity.js
@@ -2,7 +2,7 @@
 
 /**
  * Continuity store (CC-1) — the per-project "hot tier" of the Continuity
- * Contract (`.claude/plans/continuity-contract.md`). This is the thin
+ * Contract (`.claude/plans/archive/continuity-contract.md`). This is the thin
  * vertical slice that proves the WRITE→READ loop end-to-end: a wrap step
  * writes a curated **index** carrying a `Next action`, and the next
  * session's prime reads it back to offer a visible "we left off at X —
