+++ lib/wrap-steps/priming-roll.js
@@ -166,6 +166,11 @@ function _selectPointer(chunks) {
  * fences so it can sit inside a fenced surrounding markdown block
  * without conflict.
  *
+ * The caller (`run`) guarantees a resolvable pointer: it skips before
+ * rendering when the plan has zero chunks (#515), so `_selectPointer`
+ * always yields either `allDone` or a `current` — never an all-`null`
+ * pointer. There is therefore no "no headings" fallback branch here.
+ *
  * @param {{current:object|null, next:object|null, allDone:boolean}} pointer
  * @param {string} planRelPath - Plan path relative to project root, for the human reader
  * @returns {string} Body sandwiched between BEGIN/END markers (markers not included)
@@ -193,10 +198,6 @@ function _renderPointerBody(pointer, planRelPath) {
       lines.push('', '_Last chunk in this plan._');
     }
     lines.push('', `Plan: \`${planRelPath}\``);
-  } else {
-    lines.push(
-      `No \`### Chunk N: Title\` headings found in \`${planRelPath}\`.`
-    );
   }
   lines.push('');
   return lines.join('\n');
